input.c: remove unnecessary casts
This patch to Flickernoise removes a number of casts that serve no
purpose and - in the case of handle_midi_msg - clutter code that
needs some more fixing.

I didn't run the code but I checked that the assembler output is
equivalent. It's not identical because the line numbers of four
asserts changed.

There is a bit more redundancy in the form of

  ... | (foo << bar) | ...

which could be written as

  ... | foo << bar | ...

I didn't touch these since some people prefer the parentheses as
part of their coding style. If you want such things gone as well,
please let me know.

- Werner