compiler: alternative section syntax per_frame: and per_vertex:
Patch sections can now be marked with the section labels that end
in a colon and not an equal sign. This makes it clearer that they
are something very different from an assignment.

Order of sections using this new syntax is also more strictly
enforced: per_frame: must precede per_vertex: and neither of them
can appear more than once in a patch. Both sections are optional
and they can also be left empty.