compiler: rewrite grammar without precedence directives
This makes the grammar look like the ANSI C grammar in section A13
of K&R 2nd Ed., which will make it easy to add other elements with
C-like semantics.

Furthermore, not having precedences to fall back to makes sure the
parser generator will report every single ambiguity and will never
try to solve it on its own.

Last but not least, this also allows us to get rid of the TOK_NOT
pseudo-token.